Progress on codebase cleanup
