+++ src/models/bulk-upload.model.ts
@@ -1,0 +1,7 @@
+
+export interface FileUploadRequest {
+	content: string,
+	name: string,
+	size: number,
+	type: string
+}
