+++ src/models/quick-pick.model.ts
