+++ src/sftp/connection-manager.ts
@@ -1,16 +1,18 @@
 import Client from 'ssh2-sftp-client';
 import { SFTPWrapper } from 'ssh2';
-import { RemoteConfiguration } from './configuration.js';
+import { RemoteConfiguration } from '../base/configuration.js';
 import { Pool, PoolFactory } from 'lightning-pool';
-import logger from './logger.js';
 import { workspace } from 'vscode';
+import { SFTPExtension } from '../base/vscode-extension.js';
 
 export class ConnectionManager {
   private openConnections: Map<string, ResourcedPool> = new Map();
 
+  constructor(private extension: SFTPExtension) {}
+
   async createPool(openConfiguration: SFTPConnectionOpen, testSuite: boolean = false) {
-    logger.appendLineToMessages('Created connection pool for remote "' + openConfiguration.remoteName + '".');
-    const pool = new ResourcedPool(openConfiguration, testSuite);
+    this.extension.logger.appendLineToMessages('Created connection pool for remote "' + openConfiguration.remoteName + '".');
+    const pool = new ResourcedPool(this.extension, openConfiguration, testSuite);
     this.openConnections.set(openConfiguration.remoteName, pool);
   }
 
@@ -69,16 +71,16 @@ export class ResourcedPool {
   testSuitePassivePoolMaxQueue: number = -1;
   testSuitePassivePoolIdleTimeoutMillis: number = -1;
 
-  constructor(openConfiguration: SFTPConnectionOpen, testSuite: boolean = false) {
+  constructor(private extension: SFTPExtension, openConfiguration: SFTPConnectionOpen, testSuite: boolean = false) {
     this.configuration = openConfiguration.configuration;
     this.remoteName = openConfiguration.remoteName;
 
     this.passiveFactory = {
       create: async function() {
         const client = new Client("sftp-" + openConfiguration.remoteName);
-        logger.appendLineToMessages('[connection] [task] Connecting to remote SFTP "' + openConfiguration.remoteName + '"....');
+        extension.logger.appendLineToMessages('[connection] [task] Connecting to remote SFTP "' + openConfiguration.remoteName + '"....');
         const sftp = await client.connect({ ...openConfiguration.configuration });
-        logger.appendLineToMessages('[connection] [task] Connection success to remote SFTP "' + openConfiguration.remoteName + '"...');
+        extension.logger.appendLineToMessages('[connection] [task] Connection success to remote SFTP "' + openConfiguration.remoteName + '"...');
         const provider = new ConnectionProvider(client, sftp, 'passive');
   
         client.on('close', () => {
@@ -106,7 +108,7 @@ export class ResourcedPool {
         return provider;
       },
       destroy: async function(provider) {  
-        logger.appendLineToMessages('[connection] [task] Destroying connection for remote SFTP "' + openConfiguration.remoteName + '"...');
+        extension.logger.appendLineToMessages('[connection] [task] Destroying connection for remote SFTP "' + openConfiguration.remoteName + '"...');
         provider.getSFTP().end();
       },
       validate: async function(provider) {
@@ -118,7 +120,7 @@ export class ResourcedPool {
           return;
         }
 
-        logger.appendLineToMessages('[connection] [task] Validating connection for "' + openConfiguration.remoteName + '"...');
+        extension.logger.appendLineToMessages('[connection] [task] Validating connection for "' + openConfiguration.remoteName + '"...');
         const start = Date.now();
 
         await new Promise<void>((resolve, reject) => {
@@ -135,7 +137,7 @@ export class ResourcedPool {
         const end = Date.now() - start;
 
         provider.lastValidation = Date.now();
-        logger.appendLineToMessages('[connection] [task] Validated connection for "' + openConfiguration.remoteName + '" in ' + end + 'ms...');
+        extension.logger.appendLineToMessages('[connection] [task] Validated connection for "' + openConfiguration.remoteName + '" in ' + end + 'ms...');
   
         provider.getSFTP();
       }
@@ -144,9 +146,9 @@ export class ResourcedPool {
     this.heavyFactory = {
       create: async function() {
         const client = new Client("sftp-" + openConfiguration.remoteName);
-        logger.appendLineToMessages('[connection] [task] Connecting to remote SFTP "' + openConfiguration.remoteName + '"....');
+        extension.logger.appendLineToMessages('[connection] [task] Connecting to remote SFTP "' + openConfiguration.remoteName + '"....');
         const sftp = await client.connect({ ...openConfiguration.configuration });
-        logger.appendLineToMessages('[connection] [task] Connection success to remote SFTP "' + openConfiguration.remoteName + '"...');
+        extension.logger.appendLineToMessages('[connection] [task] Connection success to remote SFTP "' + openConfiguration.remoteName + '"...');
         const provider = new ConnectionProvider(client, sftp, 'heavy');
   
         client.on('close', () => {
@@ -174,7 +176,7 @@ export class ResourcedPool {
         return provider;
       },
       destroy: async function(provider) {  
-        logger.appendLineToMessages('[connection] [task] Destroying connection for remote SFTP "' + openConfiguration.remoteName + '"...');
+        extension.logger.appendLineToMessages('[connection] [task] Destroying connection for remote SFTP "' + openConfiguration.remoteName + '"...');
         provider.getSFTP().end();
       },
       validate: async function(provider) {
@@ -182,7 +184,7 @@ export class ResourcedPool {
           throw Error('SFTP already closed.');
         }
 
-        logger.appendLineToMessages('[connection] [task] Validating connection for "' + openConfiguration.remoteName + '"...');
+        extension.logger.appendLineToMessages('[connection] [task] Validating connection for "' + openConfiguration.remoteName + '"...');
 
         await new Promise<void>((resolve, reject) => {
           provider.getSFTP().stat('/', (err) => {
@@ -195,7 +197,7 @@ export class ResourcedPool {
           });
         });
 
-        logger.appendLineToMessages('[connection] [task] Validated connection for "' + openConfiguration.remoteName + '"...');
+        extension.logger.appendLineToMessages('[connection] [task] Validated connection for "' + openConfiguration.remoteName + '"...');
 
         provider.getSFTP();
       }
@@ -230,7 +232,7 @@ export class ResourcedPool {
   async reconnect() {
     this.poolPromise = new Promise(async (resolve, reject) => {
       try {
-        logger.appendLineToMessages('[pool] [task] Performing reconnection...');
+        this.extension.logger.appendLineToMessages('[pool] [task] Performing reconnection...');
         this.terminated = true;
         await this.passivePool.closeAsync(true);
         await this.heavyPool.closeAsync(true);
@@ -238,7 +240,7 @@ export class ResourcedPool {
         this.terminated = false;
         resolve();
       } catch(ex: any) {
-        logger.appendErrorToMessages('reconnect', 'Failed to reconnect: ', ex);
+        this.extension.logger.appendErrorToMessages('reconnect', 'Failed to reconnect: ', ex);
         reject(ex);
       }
     });
@@ -261,11 +263,11 @@ export class ResourcedPool {
     const passiveMaxQueue = passiveConfig.get('maxQueue', 1000000);
     const passiveIdleTimeoutMillis = passiveConfig.get('idleTimeoutMillis', 60000);
 
-    logger.appendLineToMessages('[pool-config] heavy.max = ' + heavyMax);
-    logger.appendLineToMessages('[pool-config] heavy.min = ' + heavyMin);
-    logger.appendLineToMessages('[pool-config] heavy.minIdle = ' + heavyMinIdle);
-    logger.appendLineToMessages('[pool-config] heavy.maxQueue = ' + heavyMaxQueue);
-    logger.appendLineToMessages('[pool-config] heavy.idleTimeoutMillis = ' + heavyIdleTimeoutMillis);
+    this.extension.logger.appendLineToMessages('[pool-config] heavy.max = ' + heavyMax);
+    this.extension.logger.appendLineToMessages('[pool-config] heavy.min = ' + heavyMin);
+    this.extension.logger.appendLineToMessages('[pool-config] heavy.minIdle = ' + heavyMinIdle);
+    this.extension.logger.appendLineToMessages('[pool-config] heavy.maxQueue = ' + heavyMaxQueue);
+    this.extension.logger.appendLineToMessages('[pool-config] heavy.idleTimeoutMillis = ' + heavyIdleTimeoutMillis);
 
     this.testSuiteHeavyPoolMax = heavyMax;
     this.testSuiteHeavyPoolMin = heavyMin;
@@ -273,11 +275,11 @@ export class ResourcedPool {
     this.testSuiteHeavyPoolMaxQueue = heavyMaxQueue;
     this.testSuiteHeavyPoolIdleTimeoutMillis = heavyIdleTimeoutMillis;
 
-    logger.appendLineToMessages('[pool-config] passive.max = ' + passiveMax);
-    logger.appendLineToMessages('[pool-config] passive.min = ' + passiveMin);
-    logger.appendLineToMessages('[pool-config] passive.minIdle = ' + passiveMinIdle);
-    logger.appendLineToMessages('[pool-config] passive.maxQueue = ' + passiveMaxQueue);
-    logger.appendLineToMessages('[pool-config] passive.idleTimeoutMillis = ' + passiveIdleTimeoutMillis);
+    this.extension.logger.appendLineToMessages('[pool-config] passive.max = ' + passiveMax);
+    this.extension.logger.appendLineToMessages('[pool-config] passive.min = ' + passiveMin);
+    this.extension.logger.appendLineToMessages('[pool-config] passive.minIdle = ' + passiveMinIdle);
+    this.extension.logger.appendLineToMessages('[pool-config] passive.maxQueue = ' + passiveMaxQueue);
+    this.extension.logger.appendLineToMessages('[pool-config] passive.idleTimeoutMillis = ' + passiveIdleTimeoutMillis);
 
     this.testSuitePassivePoolMax = passiveMax;
     this.testSuitePassivePoolMin = passiveMin;
@@ -285,9 +287,9 @@ export class ResourcedPool {
     this.testSuitePassivePoolMaxQueue = passiveMaxQueue;
     this.testSuitePassivePoolIdleTimeoutMillis = passiveIdleTimeoutMillis;
 
-    logger.appendLineToMessages('[pool-config] targetHost = ' + this.configuration.host);
-    logger.appendLineToMessages('[pool-config] targetPort = ' + this.configuration.port);
-    logger.appendLineToMessages('[pool-config] targetUser = ' + this.configuration.username);
+    this.extension.logger.appendLineToMessages('[pool-config] targetHost = ' + this.configuration.host);
+    this.extension.logger.appendLineToMessages('[pool-config] targetPort = ' + this.configuration.port);
+    this.extension.logger.appendLineToMessages('[pool-config] targetUser = ' + this.configuration.username);
 
     if (!testSuite) {
       this.heavyPool = new Pool(this.heavyFactory, {  
@@ -341,6 +343,3 @@ export class ConnectionProvider {
 }
 
 export type ConnectionStatus = 'OPEN' | 'OPENING' | 'CLOSED'
-
-const connectionManager = new ConnectionManager();
-export default connectionManager;
